+++ src/mma/merge/__init__.py
@@ -1,0 +1,5 @@
+"""Multitask merge package (P5)."""
+
+from mma.merge.validate_ready import validate_ready
+
+__all__ = ["validate_ready"]